requirements.txt
python-telegram-bot==13.15
gspread==5.11.1
oauth2client==4.1.3
